TA-28: Rename Pact tests
