+++ .gitignore
@@ -24,3 +24,4 @@ java-rpm-packaging/
 application-*.yml
 application-*.yaml
 application-local.properties
+/pacts/
