--- src/contractTest/java/uk/gov/hmcts/reform/divorce/casemaintenanceservice/client/DivorceCaseMaintenanceSearchForCitizen.java
@@ -1,106 +1,0 @@
-package uk.gov.hmcts.reform.divorce.casemaintenanceservice.client;
-
-import au.com.dius.pact.consumer.dsl.PactDslJsonArray;
-import au.com.dius.pact.consumer.dsl.PactDslWithProvider;
-import au.com.dius.pact.consumer.junit5.PactTestFor;
-import au.com.dius.pact.core.model.RequestResponsePact;
-import au.com.dius.pact.core.model.annotations.Pact;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.hamcrest.CoreMatchers;
-import org.json.JSONException;
-import org.junit.Before;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.MediaType;
-import uk.gov.hmcts.reform.ccd.client.CoreCaseDataApi;
-import uk.gov.hmcts.reform.ccd.client.model.CaseDataContent;
-import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
-import uk.gov.hmcts.reform.ccd.client.model.Event;
-import uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.DivorceCaseMaintenancePact;
-
-import java.io.IOException;
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-import java.util.TreeMap;
-
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.core.Is.is;
-import static org.junit.Assert.assertNotNull;
-import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.PactDslBuilderForCaseDetailsList.buildNewListOfCaseDetailsDsl;
-
-public class DivorceCaseMaintenanceSearchForCitizen  extends DivorceCaseMaintenancePact {
-
-
-    private static final String USER_ID = "123456";
-    private Map<String, Object> caseDetailsMap;
-    private CaseDataContent caseDataContent;
-    private CaseDetails caseDetails;
-    Map<String, Object> params = new TreeMap<>(String.CASE_INSENSITIVE_ORDER);
-
-    @BeforeAll
-    public void setUp() throws Exception {
-        Thread.sleep(2000);
-        caseDetailsMap = getCaseDetailsAsMap("divorce-map.json");
-        caseDataContent = CaseDataContent.builder()
-            .eventToken("someEventToken")
-            .event(
-                Event.builder()
-                    .id(createEventId)
-                    .summary(DIVORCE_CASE_SUBMISSION_EVENT_SUMMARY)
-                    .description(DIVORCE_CASE_SUBMISSION_EVENT_DESCRIPTION)
-                    .build()
-            ).data(caseDetailsMap.get("case_data"))
-            .build();
-    }
-
-    @BeforeEach
-    public void setUpEachTest() throws InterruptedException {
-        Thread.sleep(2000);
-    }
-
-    @Pact(provider = "ccdDataStoreAPI_Cases", consumer = "divorce_caseMaintenanceService")
-    RequestResponsePact searchForCitizen(PactDslWithProvider builder) {
-        params = Collections.emptyMap();
-
-        // @formatter:off
-        return builder
-            .given("A Search cases for a Citizen is requested",getCaseDataContentAsMap(caseDataContent))
-            .uponReceiving("A Search Cases for a Citizen")
-            .path("/citizens/"
-                + USER_ID
-                + "/jurisdictions/"
-                + jurisdictionId
-                + "/case-types/"
-                + caseType
-                + "/cases")
-            .method("GET")
-            .query("")
-            .headers(HttpHeaders.AUTHORIZATION, SOME_AUTHORIZATION_TOKEN, SERVICE_AUTHORIZATION,
-                SOME_SERVICE_AUTHORIZATION_TOKEN)
-            .matchHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-            .willRespondWith()
-            .matchHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-            .status(200)
-            .body(
-                buildNewListOfCaseDetailsDsl(20000000L))
-            .toPact();
-    }
-
-    @Test
-    @PactTestFor(pactMethod = "searchForCitizen")
-    public void searchForCitizen() throws IOException, JSONException {
-        final Map<String, String> searchCriteria = Collections.EMPTY_MAP;
-
-        List<CaseDetails> caseDetailsList = coreCaseDataApi.searchForCitizen(SOME_AUTHORIZATION_TOKEN,
-            SOME_SERVICE_AUTHORIZATION_TOKEN, USER_ID, jurisdictionId,
-            caseType, searchCriteria);
-
-        assertNotNull(caseDetailsList);
-
-    }
-}
