+++ src/contractTest/java/uk/gov/hmcts/reform/divorce/casemaintenanceservice/client/ccd/ReadForCitizenConsumerTest.java
@@ -1,12 +1,13 @@
-package uk.gov.hmcts.reform.divorce.casemaintenanceservice.client;
+package uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.ccd;
 
 import au.com.dius.pact.consumer.dsl.PactDslWithProvider;
 import au.com.dius.pact.consumer.junit5.PactTestFor;
 import au.com.dius.pact.core.model.RequestResponsePact;
 import au.com.dius.pact.core.model.annotations.Pact;
 import org.apache.http.client.fluent.Executor;
+import org.hamcrest.core.Is;
 import org.json.JSONException;
-import org.junit.jupiter.api.AfterEach;
+import org.junit.After;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpHeaders;
@@ -14,27 +15,24 @@ import org.springframework.http.MediaType;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDataContent;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.ccd.client.model.Event;
-import uk.gov.hmcts.reform.ccd.client.model.StartEventResponse;
-import uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.DivorceCaseMaintenancePact;
+import uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.CcdConsumerTestBase;
 
 import java.io.IOException;
 import java.util.Map;
 
-import static junit.framework.TestCase.assertNotNull;
-import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.PactDslBuilderForCaseDetailsList.buildStartEventResponseWithEmptyCaseDetails;
+import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.AssertionHelper.assertCaseDetails;
+import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.PactDslBuilderForCaseDetailsList.buildCaseDetailsDsl;
 
-public class DivorceCaseMaintenanceStartForCitizen extends DivorceCaseMaintenancePact {
 
-    public static final String EVENT_ID = "eventId";
+public class ReadForCitizenConsumerTest extends CcdConsumerTestBase {
 
     private Map<String, Object> caseDetailsMap;
     private CaseDataContent caseDataContent;
 
-
     @BeforeAll
     public void setUp() throws Exception {
+        Thread.sleep(2000);
         caseDetailsMap = getCaseDetailsAsMap("divorce-map.json");
         caseDataContent = CaseDataContent.builder()
             .eventToken("someEventToken")
@@ -49,17 +47,19 @@ public class DivorceCaseMaintenanceStartForCitizen extends DivorceCaseMaintenanc
     }
 
     @Pact(provider = "ccdDataStoreAPI_Cases", consumer = "divorce_caseMaintenanceService")
-    RequestResponsePact startForCitizen(PactDslWithProvider builder) {
+    RequestResponsePact readForCitizen(PactDslWithProvider builder) {
         // @formatter:off
         return builder
-            .given("A Start for a Citizen is requested", getCaseDataContentAsMap(caseDataContent))
-            .uponReceiving("A Start for a Citizen")
-            .path("/citizens/" + USER_ID + "/jurisdictions/"
-                + jurisdictionId + "/case-types/"
+            .given("A Read for a Citizen is requested", getCaseDataContentAsMap(caseDataContent))
+            .uponReceiving("A Read For a Citizen")
+            .path("/citizens/"
+                + USER_ID
+                + "/jurisdictions/"
+                + jurisdictionId
+                + "/case-types/"
                 + caseType
-                + "/event-triggers/"
-                + createEventId
-                + "/token")
+                + "/cases/"
+                +  CASE_ID)
             .method("GET")
             .headers(HttpHeaders.AUTHORIZATION, SOME_AUTHORIZATION_TOKEN, SERVICE_AUTHORIZATION,
                 SOME_SERVICE_AUTHORIZATION_TOKEN)
@@ -67,37 +67,27 @@ public class DivorceCaseMaintenanceStartForCitizen extends DivorceCaseMaintenanc
             .willRespondWith()
             .matchHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
             .status(200)
-            .body(buildStartEventResponseWithEmptyCaseDetails(createEventId))
+            .body(buildCaseDetailsDsl(CASE_ID))
             .toPact();
     }
 
     @Test
-    @PactTestFor(pactMethod = "startForCitizen")
-    public void verifyStartForCitizen() throws IOException, JSONException {
+    @PactTestFor(pactMethod = "readForCitizen")
+    public void verifyReadForCitizen() throws IOException, JSONException {
 
-        StartEventResponse startEventResponse = coreCaseDataApi.startForCitizen(SOME_AUTHORIZATION_TOKEN,
+        CaseDetails caseDetailsReponse = coreCaseDataApi.readForCitizen(SOME_AUTHORIZATION_TOKEN,
             SOME_SERVICE_AUTHORIZATION_TOKEN, USER_ID, jurisdictionId,
-            caseType, createEventId);
+            caseType,String.valueOf(CASE_ID));
 
-        assertThat(startEventResponse.getEventId(), equalTo(createEventId));
+        assertThat(caseDetailsReponse.getId(), Is.is(CASE_ID));
+        assertThat(caseDetailsReponse.getJurisdiction(), Is.is("DIVORCE"));
 
-        CaseDetails caseDetails = startEventResponse.getCaseDetails();
-
-        assertNotNull(startEventResponse.getCaseDetails());
+        assertCaseDetails(caseDetailsReponse);
 
     }
 
-
-    @Override
-    protected Map<String, Object> getCaseDataContentAsMap(CaseDataContent caseDataContent) throws JSONException {
-        Map<String, Object> caseDataContentMap = super.getCaseDataContentAsMap(caseDataContent);
-        caseDataContentMap.put(EVENT_ID, createEventId);
-        return caseDataContentMap;
-    }
-
-    @AfterEach
+    @After
     void teardown() {
         Executor.closeIdleConnections();
     }
-
 }
