+++ src/contractTest/java/uk/gov/hmcts/reform/divorce/casemaintenanceservice/client/ccd/SearchForCitizenConsumerTest.java
@@ -1,31 +1,39 @@
-package uk.gov.hmcts.reform.divorce.casemaintenanceservice.client;
+package uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.ccd;
 
 import au.com.dius.pact.consumer.dsl.PactDslWithProvider;
 import au.com.dius.pact.consumer.junit5.PactTestFor;
 import au.com.dius.pact.core.model.RequestResponsePact;
 import au.com.dius.pact.core.model.annotations.Pact;
-import org.apache.http.client.fluent.Executor;
 import org.json.JSONException;
-import org.junit.After;
 import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDataContent;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.ccd.client.model.Event;
-import uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.DivorceCaseMaintenancePact;
+import uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.CcdConsumerTestBase;
 
 import java.io.IOException;
+import java.util.Collections;
+import java.util.List;
 import java.util.Map;
+import java.util.TreeMap;
 
-import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.AssertionHelper.assertCaseDetails;
-import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.PactDslBuilderForCaseDetailsList.buildCaseDetailsDsl;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.core.Is.is;
+import static org.junit.Assert.assertNotNull;
+import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.PactDslBuilderForCaseDetailsList.buildNewListOfCaseDetailsDsl;
 
-public class DivorceCaseMaintenanceReadForCaseWorker  extends DivorceCaseMaintenancePact {
+public class SearchForCitizenConsumerTest extends CcdConsumerTestBase {
 
+
+    private static final String USER_ID = "123456";
     private Map<String, Object> caseDetailsMap;
     private CaseDataContent caseDataContent;
+    private CaseDetails caseDetails;
+    Map<String, Object> params = new TreeMap<>(String.CASE_INSENSITIVE_ORDER);
 
     @BeforeAll
     public void setUp() throws Exception {
@@ -43,46 +51,49 @@ public class DivorceCaseMaintenanceReadForCaseWorker  extends DivorceCaseMainten
             .build();
     }
 
+    @BeforeEach
+    public void setUpEachTest() throws InterruptedException {
+        Thread.sleep(2000);
+    }
+
     @Pact(provider = "ccdDataStoreAPI_Cases", consumer = "divorce_caseMaintenanceService")
-    RequestResponsePact readForCaseDetails(PactDslWithProvider builder) {
+    RequestResponsePact searchForCitizen(PactDslWithProvider builder) {
+        params = Collections.emptyMap();
+
         // @formatter:off
         return builder
-            .given("A Read for a Caseworker is requested", getCaseDataContentAsMap(caseDataContent))
-            .uponReceiving("A Read For a Caseworker")
-            .path("/caseworkers/"
+            .given("A Search cases for a Citizen is requested",getCaseDataContentAsMap(caseDataContent))
+            .uponReceiving("A Search Cases for a Citizen")
+            .path("/citizens/"
                 + USER_ID
                 + "/jurisdictions/"
                 + jurisdictionId
                 + "/case-types/"
                 + caseType
-                + "/cases/"
-                +  CASE_ID)
+                + "/cases")
             .method("GET")
+            .query("")
             .headers(HttpHeaders.AUTHORIZATION, SOME_AUTHORIZATION_TOKEN, SERVICE_AUTHORIZATION,
                 SOME_SERVICE_AUTHORIZATION_TOKEN)
             .matchHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
             .willRespondWith()
             .matchHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
             .status(200)
-            .body(buildCaseDetailsDsl(CASE_ID))
+            .body(
+                buildNewListOfCaseDetailsDsl(20000000L))
             .toPact();
     }
 
     @Test
-    @PactTestFor(pactMethod = "readForCaseDetails")
-    public void verifyReadForCaseDetails() throws IOException, JSONException {
+    @PactTestFor(pactMethod = "searchForCitizen")
+    public void searchForCitizen() throws IOException, JSONException {
+        final Map<String, String> searchCriteria = Collections.EMPTY_MAP;
 
-        CaseDetails caseDetailsReponse = coreCaseDataApi.readForCaseWorker(SOME_AUTHORIZATION_TOKEN,
+        List<CaseDetails> caseDetailsList = coreCaseDataApi.searchForCitizen(SOME_AUTHORIZATION_TOKEN,
             SOME_SERVICE_AUTHORIZATION_TOKEN, USER_ID, jurisdictionId,
-            caseType,CASE_ID.toString());
+            caseType, searchCriteria);
 
-        assertCaseDetails(caseDetailsReponse);
+        assertNotNull(caseDetailsList);
 
     }
-
-    @After
-    void teardown() {
-        Executor.closeIdleConnections();
-    }
-
 }
