+++ src/contractTest/java/uk/gov/hmcts/reform/divorce/casemaintenanceservice/client/ccd/StartEventForCaseWorkerConsumerTest.java
@@ -1,41 +1,37 @@
-package uk.gov.hmcts.reform.divorce.casemaintenanceservice.client;
+package uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.ccd;
 
 import au.com.dius.pact.consumer.dsl.PactDslWithProvider;
 import au.com.dius.pact.consumer.junit5.PactTestFor;
 import au.com.dius.pact.core.model.RequestResponsePact;
 import au.com.dius.pact.core.model.annotations.Pact;
-import org.apache.http.client.fluent.Executor;
 import org.json.JSONException;
-import org.junit.After;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDataContent;
-import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.ccd.client.model.Event;
-import uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.DivorceCaseMaintenancePact;
-import uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.PactDslFixtureHelper;
+import uk.gov.hmcts.reform.ccd.client.model.StartEventResponse;
+import uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.CcdConsumerTestBase;
 
 import java.util.Map;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.AssertionHelper.assertCaseDetails;
-import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.ObjectMapperTestUtil.convertObjectToJsonString;
-import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.PactDslBuilderForCaseDetailsList.buildCaseDetailsDsl;
-import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.PactDslFixtureHelper.getCaseDataContent;
+import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.PactDslBuilderForCaseDetailsList.buildStartEventReponse;
 
-public class DivorceCaseMaintenanceSubmitEventForCaseWorker extends DivorceCaseMaintenancePact {
+public class StartEventForCaseWorkerConsumerTest extends CcdConsumerTestBase {
 
     public static final String HWF_APPLICATION_ACCEPTED = "hwfApplicationAccepted";
+
     private Map<String, Object> caseDetailsMap;
     private CaseDataContent caseDataContent;
 
+
     @BeforeAll
     public void setUp() throws Exception {
-
         Thread.sleep(2000);
         caseDetailsMap = getCaseDetailsAsMap("divorce-map.json");
         caseDataContent = CaseDataContent.builder()
@@ -50,43 +46,44 @@ public class DivorceCaseMaintenanceSubmitEventForCaseWorker extends DivorceCaseM
             .build();
     }
 
+    @BeforeEach
+    public void setUpEachTest() throws InterruptedException {
+        Thread.sleep(2000);
+    }
 
     @Pact(provider = "ccdDataStoreAPI_Cases", consumer = "divorce_caseMaintenanceService")
-    RequestResponsePact submitEventForCaseWorker(PactDslWithProvider builder) throws Exception {
+    RequestResponsePact startEventForCaseWorker(PactDslWithProvider builder) {
         // @formatter:off
         return builder
-            .given("A Submit Event for a Caseworker is requested", getCaseDataContentAsMap(caseDataContent))
-            .uponReceiving("A Submit Event for a Caseworker")
-            .path("/caseworkers/" + USER_ID
-                + "/jurisdictions/" + jurisdictionId
-                + "/case-types/" + caseType
-                + "/cases/" + CASE_ID
-                + "/events"
-            )
-            .query("ignore-warning=true")
-            .method("POST")
+            .given("A Start Event for a Caseworker is  requested",getCaseDataContentAsMap(caseDataContent))
+            .uponReceiving("A Start Event for a Caseworker")
+            .path("/caseworkers/" + USER_ID + "/jurisdictions/"
+                + jurisdictionId + "/case-types/"
+                + caseType
+                + "/cases/"
+                +  CASE_ID
+                + "/event-triggers/"
+                + HWF_APPLICATION_ACCEPTED
+                + "/token")
+            .method("GET")
             .headers(HttpHeaders.AUTHORIZATION, SOME_AUTHORIZATION_TOKEN, SERVICE_AUTHORIZATION, SOME_SERVICE_AUTHORIZATION_TOKEN)
-            .body(convertObjectToJsonString(getCaseDataContent(HWF_APPLICATION_ACCEPTED)))
             .matchHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
             .willRespondWith()
-            .status(201)
-            .body(buildCaseDetailsDsl(CASE_ID))
             .matchHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
+            .status(200)
+            .body(buildStartEventReponse(HWF_APPLICATION_ACCEPTED))
             .toPact();
     }
 
     @Test
-    @PactTestFor(pactMethod = "submitEventForCaseWorker")
-    public void verifySubmitEventForCaseworker() throws Exception {
+    @PactTestFor(pactMethod = "startEventForCaseWorker")
+    public void verifyStartEventForCaseworker() throws JSONException {
 
-        caseDataContent = PactDslFixtureHelper.getCaseDataContent(HWF_APPLICATION_ACCEPTED);
+        final StartEventResponse startEventResponse = coreCaseDataApi.startEventForCaseWorker(SOME_AUTHORIZATION_TOKEN,
+            SOME_SERVICE_AUTHORIZATION_TOKEN, USER_ID, jurisdictionId,caseType,CASE_ID.toString(),HWF_APPLICATION_ACCEPTED);
 
-        final CaseDetails caseDetails = coreCaseDataApi.submitEventForCaseWorker(SOME_AUTHORIZATION_TOKEN,
-            SOME_SERVICE_AUTHORIZATION_TOKEN, USER_ID, jurisdictionId, caseType, CASE_ID.toString(), true, caseDataContent);
-
-        assertThat(caseDetails.getId(), is(CASE_ID));
-        assertThat(caseDetails.getJurisdiction(), is("DIVORCE"));
-        assertCaseDetails(caseDetails);
+        assertThat(startEventResponse.getEventId(), is(HWF_APPLICATION_ACCEPTED));
+        assertCaseDetails(startEventResponse.getCaseDetails());
     }
 
     @Override
@@ -94,11 +91,6 @@ public class DivorceCaseMaintenanceSubmitEventForCaseWorker extends DivorceCaseM
         Map<String, Object> caseDataContentMap = super.getCaseDataContentAsMap(caseDataContent);
         caseDataContentMap.put(EVENT_ID, HWF_APPLICATION_ACCEPTED);
         return caseDataContentMap;
-    }
-
-    @After
-    public void teardown() {
-        Executor.closeIdleConnections();
     }
 
 }
