+++ src/contractTest/java/uk/gov/hmcts/reform/divorce/casemaintenanceservice/client/ccd/StartEventForCitizenConsumerTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.divorce.casemaintenanceservice.client;
+package uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.ccd;
 
 import au.com.dius.pact.consumer.dsl.PactDslWithProvider;
 import au.com.dius.pact.consumer.junit5.PactTestFor;
@@ -6,29 +6,25 @@ import au.com.dius.pact.core.model.RequestResponsePact;
 import au.com.dius.pact.core.model.annotations.Pact;
 import org.json.JSONException;
 import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
-import uk.gov.hmcts.reform.ccd.client.CoreCaseDataApi;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDataContent;
 import uk.gov.hmcts.reform.ccd.client.model.Event;
 import uk.gov.hmcts.reform.ccd.client.model.StartEventResponse;
-import uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.DivorceCaseMaintenancePact;
+import uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.CcdConsumerTestBase;
 
-import java.util.HashMap;
 import java.util.Map;
 
+import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.AssertionHelper.assertCaseDetails;
 import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.PactDslBuilderForCaseDetailsList.buildStartEventReponse;
 
-public class DivorceCaseMaintenanceStartEventForCaseWorker  extends DivorceCaseMaintenancePact {
+public class StartEventForCitizenConsumerTest extends CcdConsumerTestBase {
 
-    public static final String HWF_APPLICATION_ACCEPTED = "hwfApplicationAccepted";
+    public static final String PAYMENT_REFERENCE_GENERATED = "paymentReferenceGenerated";
 
     private Map<String, Object> caseDetailsMap;
     private CaseDataContent caseDataContent;
@@ -50,50 +46,50 @@ public class DivorceCaseMaintenanceStartEventForCaseWorker  extends DivorceCaseM
             .build();
     }
 
-    @BeforeEach
-    public void setUpEachTest() throws InterruptedException {
-        Thread.sleep(2000);
-    }
 
     @Pact(provider = "ccdDataStoreAPI_Cases", consumer = "divorce_caseMaintenanceService")
-    RequestResponsePact startEventForCaseWorker(PactDslWithProvider builder) {
+    RequestResponsePact startEventForCitizen(PactDslWithProvider builder) {
         // @formatter:off
         return builder
-            .given("A Start Event for a Caseworker is  requested",getCaseDataContentAsMap(caseDataContent))
-            .uponReceiving("A Start Event for a Caseworker")
-            .path("/caseworkers/" + USER_ID + "/jurisdictions/"
+            .given("A Start Event for a Citizen is requested", getCaseDataContentAsMap(caseDataContent))
+            .uponReceiving("A Start Event a Citizen")
+            .path("/citizens/" + USER_ID + "/jurisdictions/"
                 + jurisdictionId + "/case-types/"
                 + caseType
-                + "/cases/"
-                +  CASE_ID
+                + "/cases/" + CASE_ID
                 + "/event-triggers/"
-                + HWF_APPLICATION_ACCEPTED
+                + PAYMENT_REFERENCE_GENERATED
                 + "/token")
             .method("GET")
-            .headers(HttpHeaders.AUTHORIZATION, SOME_AUTHORIZATION_TOKEN, SERVICE_AUTHORIZATION, SOME_SERVICE_AUTHORIZATION_TOKEN)
+            .headers(HttpHeaders.AUTHORIZATION, SOME_AUTHORIZATION_TOKEN, SERVICE_AUTHORIZATION,
+                SOME_SERVICE_AUTHORIZATION_TOKEN)
             .matchHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
             .willRespondWith()
             .matchHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
             .status(200)
-            .body(buildStartEventReponse(HWF_APPLICATION_ACCEPTED))
+            .body(buildStartEventReponse(PAYMENT_REFERENCE_GENERATED))
             .toPact();
     }
 
     @Test
-    @PactTestFor(pactMethod = "startEventForCaseWorker")
-    public void verifyStartEventForCaseworker() throws JSONException {
+    @PactTestFor(pactMethod = "startEventForCitizen")
+    public void verifyStartEventForCitizen() throws JSONException {
 
-        final StartEventResponse startEventResponse = coreCaseDataApi.startEventForCaseWorker(SOME_AUTHORIZATION_TOKEN,
-            SOME_SERVICE_AUTHORIZATION_TOKEN, USER_ID, jurisdictionId,caseType,CASE_ID.toString(),HWF_APPLICATION_ACCEPTED);
+        StartEventResponse startEventResponse = coreCaseDataApi.startEventForCitizen(SOME_AUTHORIZATION_TOKEN,
+            SOME_SERVICE_AUTHORIZATION_TOKEN, USER_ID.toString(), jurisdictionId,
+            caseType, CASE_ID.toString(), PAYMENT_REFERENCE_GENERATED);
 
-        assertThat(startEventResponse.getEventId(), is(HWF_APPLICATION_ACCEPTED));
+        assertThat(startEventResponse.getEventId(), equalTo(PAYMENT_REFERENCE_GENERATED));
+        assertThat(startEventResponse.getCaseDetails().getId(), is(2000L));
+
         assertCaseDetails(startEventResponse.getCaseDetails());
+
     }
 
     @Override
     protected Map<String, Object> getCaseDataContentAsMap(CaseDataContent caseDataContent) throws JSONException {
         Map<String, Object> caseDataContentMap = super.getCaseDataContentAsMap(caseDataContent);
-        caseDataContentMap.put(EVENT_ID, HWF_APPLICATION_ACCEPTED);
+        caseDataContentMap.put(EVENT_ID, PAYMENT_REFERENCE_GENERATED);
         return caseDataContentMap;
     }
