+++ src/contractTest/java/uk/gov/hmcts/reform/divorce/casemaintenanceservice/client/ccd/StartForCaseWorkerConsumerTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.divorce.casemaintenanceservice.client;
+package uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.ccd;
 
 import au.com.dius.pact.consumer.dsl.PactDslWithProvider;
 import au.com.dius.pact.consumer.junit5.PactTestFor;
@@ -6,26 +6,26 @@ import au.com.dius.pact.core.model.RequestResponsePact;
 import au.com.dius.pact.core.model.annotations.Pact;
 import org.json.JSONException;
 import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDataContent;
-import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.ccd.client.model.Event;
 import uk.gov.hmcts.reform.ccd.client.model.StartEventResponse;
-import uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.DivorceCaseMaintenancePact;
+import uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.CcdConsumerTestBase;
 
+import java.io.IOException;
 import java.util.Map;
 
+import static junit.framework.TestCase.assertNotNull;
 import static org.hamcrest.CoreMatchers.equalTo;
-import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.AssertionHelper.assertCaseDetails;
-import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.PactDslBuilderForCaseDetailsList.buildStartEventReponse;
+import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.PactDslBuilderForCaseDetailsList.buildStartEventResponseWithEmptyCaseDetails;
 
-public class DivorceCaseMaintenanceStartEventForCitizen extends DivorceCaseMaintenancePact {
+public class StartForCaseWorkerConsumerTest extends CcdConsumerTestBase {
 
-    public static final String PAYMENT_REFERENCE_GENERATED = "paymentReferenceGenerated";
+    public static final String EVENT_ID = "eventId";
 
     private Map<String, Object> caseDetailsMap;
     private CaseDataContent caseDataContent;
@@ -47,19 +47,22 @@ public class DivorceCaseMaintenanceStartEventForCitizen extends DivorceCaseMaint
             .build();
     }
 
+    @BeforeEach
+    public void setUpEachTest() throws InterruptedException {
+        Thread.sleep(2000);
+    }
 
     @Pact(provider = "ccdDataStoreAPI_Cases", consumer = "divorce_caseMaintenanceService")
-    RequestResponsePact startEventForCitizen(PactDslWithProvider builder) {
+    RequestResponsePact startForCaseWorker(PactDslWithProvider builder) {
         // @formatter:off
         return builder
-            .given("A Start Event for a Citizen is requested", getCaseDataContentAsMap(caseDataContent))
-            .uponReceiving("A Start Event a Citizen")
-            .path("/citizens/" + USER_ID + "/jurisdictions/"
+            .given("A Start for a Caseworker is requested", getCaseDataContentAsMap(caseDataContent))
+            .uponReceiving("A Start for a Caseworker")
+            .path("/caseworkers/" + USER_ID + "/jurisdictions/"
                 + jurisdictionId + "/case-types/"
                 + caseType
-                + "/cases/" + CASE_ID
                 + "/event-triggers/"
-                + PAYMENT_REFERENCE_GENERATED
+                + createEventId
                 + "/token")
             .method("GET")
             .headers(HttpHeaders.AUTHORIZATION, SOME_AUTHORIZATION_TOKEN, SERVICE_AUTHORIZATION,
@@ -68,29 +71,28 @@ public class DivorceCaseMaintenanceStartEventForCitizen extends DivorceCaseMaint
             .willRespondWith()
             .matchHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
             .status(200)
-            .body(buildStartEventReponse(PAYMENT_REFERENCE_GENERATED))
+            .body(buildStartEventResponseWithEmptyCaseDetails(createEventId))
             .toPact();
     }
 
     @Test
-    @PactTestFor(pactMethod = "startEventForCitizen")
-    public void verifyStartEventForCitizen() throws JSONException {
+    @PactTestFor(pactMethod = "startForCaseWorker")
+    public void verifyStartEventForCaseworker() throws IOException, JSONException {
 
-        StartEventResponse startEventResponse = coreCaseDataApi.startEventForCitizen(SOME_AUTHORIZATION_TOKEN,
-            SOME_SERVICE_AUTHORIZATION_TOKEN, USER_ID.toString(), jurisdictionId,
-            caseType, CASE_ID.toString(), PAYMENT_REFERENCE_GENERATED);
+        StartEventResponse startEventResponse = coreCaseDataApi.startForCaseworker(SOME_AUTHORIZATION_TOKEN,
+            SOME_SERVICE_AUTHORIZATION_TOKEN, USER_ID, jurisdictionId,
+            caseType, createEventId);
 
-        assertThat(startEventResponse.getEventId(), equalTo(PAYMENT_REFERENCE_GENERATED));
-        assertThat(startEventResponse.getCaseDetails().getId(), is(2000L));
+        assertThat(startEventResponse.getEventId(), equalTo(createEventId));
 
-        assertCaseDetails(startEventResponse.getCaseDetails());
+        assertNotNull(startEventResponse.getCaseDetails());
 
     }
 
     @Override
     protected Map<String, Object> getCaseDataContentAsMap(CaseDataContent caseDataContent) throws JSONException {
         Map<String, Object> caseDataContentMap = super.getCaseDataContentAsMap(caseDataContent);
-        caseDataContentMap.put(EVENT_ID, PAYMENT_REFERENCE_GENERATED);
+        caseDataContentMap.put(EVENT_ID, createEventId);
         return caseDataContentMap;
     }
