+++ src/contractTest/java/uk/gov/hmcts/reform/divorce/casemaintenanceservice/client/ccd/SubmitEventForCaseWorkerConsumerTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.divorce.casemaintenanceservice.client;
+package uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.ccd;
 
 import au.com.dius.pact.consumer.dsl.PactDslWithProvider;
 import au.com.dius.pact.consumer.junit5.PactTestFor;
@@ -8,21 +8,16 @@ import org.apache.http.client.fluent.Executor;
 import org.json.JSONException;
 import org.junit.After;
 import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
-import uk.gov.hmcts.reform.ccd.client.CoreCaseDataApi;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDataContent;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.ccd.client.model.Event;
-import uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.DivorceCaseMaintenancePact;
+import uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.CcdConsumerTestBase;
 import uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.PactDslFixtureHelper;
 
 import java.util.Map;
-import java.util.TreeMap;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -31,9 +26,7 @@ import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.Obj
 import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.PactDslBuilderForCaseDetailsList.buildCaseDetailsDsl;
 import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.PactDslFixtureHelper.getCaseDataContent;
 
-
-public class DivorceCaseMaintenanceSubmitEventForCitizen extends DivorceCaseMaintenancePact {
-
+public class SubmitEventForCaseWorkerConsumerTest extends CcdConsumerTestBase {
 
     public static final String HWF_APPLICATION_ACCEPTED = "hwfApplicationAccepted";
     private Map<String, Object> caseDetailsMap;
@@ -57,48 +50,38 @@ public class DivorceCaseMaintenanceSubmitEventForCitizen extends DivorceCaseMain
     }
 
 
-    @After
-    public void teardown() {
-        Executor.closeIdleConnections();
-    }
-
     @Pact(provider = "ccdDataStoreAPI_Cases", consumer = "divorce_caseMaintenanceService")
-    RequestResponsePact submitEventForCitizen(PactDslWithProvider builder) throws Exception {
+    RequestResponsePact submitEventForCaseWorker(PactDslWithProvider builder) throws Exception {
         // @formatter:off
         return builder
-            .given("A Submit Event for a Citizen is requested", getCaseDataContentAsMap(caseDataContent))
-            .uponReceiving("A Submit Event for a Citizen")
-            .path("/citizens/"
-                + USER_ID
-                + "/jurisdictions/"
-                + jurisdictionId
-                + "/case-types/"
-                + caseType
-                + "/cases/"
-                +  CASE_ID
-                + "/events")
+            .given("A Submit Event for a Caseworker is requested", getCaseDataContentAsMap(caseDataContent))
+            .uponReceiving("A Submit Event for a Caseworker")
+            .path("/caseworkers/" + USER_ID
+                + "/jurisdictions/" + jurisdictionId
+                + "/case-types/" + caseType
+                + "/cases/" + CASE_ID
+                + "/events"
+            )
             .query("ignore-warning=true")
             .method("POST")
+            .headers(HttpHeaders.AUTHORIZATION, SOME_AUTHORIZATION_TOKEN, SERVICE_AUTHORIZATION, SOME_SERVICE_AUTHORIZATION_TOKEN)
             .body(convertObjectToJsonString(getCaseDataContent(HWF_APPLICATION_ACCEPTED)))
-            .headers(HttpHeaders.AUTHORIZATION, SOME_AUTHORIZATION_TOKEN, SERVICE_AUTHORIZATION,
-                SOME_SERVICE_AUTHORIZATION_TOKEN)
             .matchHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
             .willRespondWith()
-            .matchHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
             .status(201)
             .body(buildCaseDetailsDsl(CASE_ID))
+            .matchHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
             .toPact();
     }
 
     @Test
-    @PactTestFor(pactMethod = "submitEventForCitizen")
-    public void verifySubmitEventForCitizen() throws Exception {
+    @PactTestFor(pactMethod = "submitEventForCaseWorker")
+    public void verifySubmitEventForCaseworker() throws Exception {
 
         caseDataContent = PactDslFixtureHelper.getCaseDataContent(HWF_APPLICATION_ACCEPTED);
 
-        CaseDetails caseDetails = coreCaseDataApi.submitEventForCitizen(SOME_AUTHORIZATION_TOKEN,
-            SOME_SERVICE_AUTHORIZATION_TOKEN, USER_ID, jurisdictionId,
-            caseType,CASE_ID.toString(),true,caseDataContent);
+        final CaseDetails caseDetails = coreCaseDataApi.submitEventForCaseWorker(SOME_AUTHORIZATION_TOKEN,
+            SOME_SERVICE_AUTHORIZATION_TOKEN, USER_ID, jurisdictionId, caseType, CASE_ID.toString(), true, caseDataContent);
 
         assertThat(caseDetails.getId(), is(CASE_ID));
         assertThat(caseDetails.getJurisdiction(), is("DIVORCE"));
@@ -108,8 +91,13 @@ public class DivorceCaseMaintenanceSubmitEventForCitizen extends DivorceCaseMain
     @Override
     protected Map<String, Object> getCaseDataContentAsMap(CaseDataContent caseDataContent) throws JSONException {
         Map<String, Object> caseDataContentMap = super.getCaseDataContentAsMap(caseDataContent);
-        caseDataContentMap.put(EVENT_ID, caseDataContent.getEvent().getId());
+        caseDataContentMap.put(EVENT_ID, HWF_APPLICATION_ACCEPTED);
         return caseDataContentMap;
+    }
+
+    @After
+    public void teardown() {
+        Executor.closeIdleConnections();
     }
 
 }
