+++ src/contractTest/java/uk/gov/hmcts/reform/divorce/casemaintenanceservice/client/ccd/SubmitEventForCitizenConsumerTest.java
@@ -1,46 +1,44 @@
-package uk.gov.hmcts.reform.divorce.casemaintenanceservice.client;
+package uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.ccd;
 
 import au.com.dius.pact.consumer.dsl.PactDslWithProvider;
 import au.com.dius.pact.consumer.junit5.PactTestFor;
 import au.com.dius.pact.core.model.RequestResponsePact;
 import au.com.dius.pact.core.model.annotations.Pact;
-import org.apache.http.HttpStatus;
 import org.apache.http.client.fluent.Executor;
 import org.json.JSONException;
 import org.junit.After;
 import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
-import uk.gov.hmcts.reform.ccd.client.CoreCaseDataApi;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDataContent;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.ccd.client.model.Event;
-import uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.DivorceCaseMaintenancePact;
+import uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.CcdConsumerTestBase;
+import uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.PactDslFixtureHelper;
 
 import java.util.Map;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.MatcherAssert.assertThat;
 import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.AssertionHelper.assertCaseDetails;
 import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.ObjectMapperTestUtil.convertObjectToJsonString;
 import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.PactDslBuilderForCaseDetailsList.buildCaseDetailsDsl;
 import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.PactDslFixtureHelper.getCaseDataContent;
-import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.PactDslFixtureHelper.getCaseDataContentWithPath;
 
-public class DivorceCaseMaintenanceSubmitForCaseWorker extends DivorceCaseMaintenancePact {
 
+public class SubmitEventForCitizenConsumerTest extends CcdConsumerTestBase {
+
+
+    public static final String HWF_APPLICATION_ACCEPTED = "hwfApplicationAccepted";
+    private Map<String, Object> caseDetailsMap;
     private CaseDataContent caseDataContent;
-
-    private static final String VALID_PAYLOAD_PATH = "json/divorce-submit.json";
 
     @BeforeAll
     public void setUp() throws Exception {
 
         Thread.sleep(2000);
+        caseDetailsMap = getCaseDetailsAsMap("divorce-map.json");
         caseDataContent = CaseDataContent.builder()
             .eventToken("someEventToken")
             .event(
@@ -49,62 +47,63 @@ public class DivorceCaseMaintenanceSubmitForCaseWorker extends DivorceCaseMainte
                     .summary(DIVORCE_CASE_SUBMISSION_EVENT_SUMMARY)
                     .description(DIVORCE_CASE_SUBMISSION_EVENT_DESCRIPTION)
                     .build()
-            )
+            ).data(caseDetailsMap.get("case_data"))
             .build();
     }
 
+
     @After
-    void teardown() {
+    public void teardown() {
         Executor.closeIdleConnections();
     }
 
     @Pact(provider = "ccdDataStoreAPI_Cases", consumer = "divorce_caseMaintenanceService")
-    RequestResponsePact submitCaseWorkerDetails(PactDslWithProvider builder) throws Exception {
+    RequestResponsePact submitEventForCitizen(PactDslWithProvider builder) throws Exception {
         // @formatter:off
         return builder
-            .given("A Submit for a Caseworker is requested" ,  getCaseDataContentAsMap(caseDataContent))
-            .uponReceiving("A Submit For a Caseworker")
-            .path("/caseworkers/"
+            .given("A Submit Event for a Citizen is requested", getCaseDataContentAsMap(caseDataContent))
+            .uponReceiving("A Submit Event for a Citizen")
+            .path("/citizens/"
                 + USER_ID
                 + "/jurisdictions/"
                 + jurisdictionId
                 + "/case-types/"
                 + caseType
-                + "/cases")
+                + "/cases/"
+                +  CASE_ID
+                + "/events")
             .query("ignore-warning=true")
             .method("POST")
-            .headers(HttpHeaders.AUTHORIZATION, SOME_AUTHORIZATION_TOKEN, SERVICE_AUTHORIZATION, SOME_SERVICE_AUTHORIZATION_TOKEN)
-            .body(convertObjectToJsonString(getCaseDataContentWithPath(createEventId, VALID_PAYLOAD_PATH)))
+            .body(convertObjectToJsonString(getCaseDataContent(HWF_APPLICATION_ACCEPTED)))
+            .headers(HttpHeaders.AUTHORIZATION, SOME_AUTHORIZATION_TOKEN, SERVICE_AUTHORIZATION,
+                SOME_SERVICE_AUTHORIZATION_TOKEN)
             .matchHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
             .willRespondWith()
-            .status(HttpStatus.SC_CREATED)
-            .body(buildCaseDetailsDsl(CASE_ID))
             .matchHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
+            .status(201)
+            .body(buildCaseDetailsDsl(CASE_ID))
             .toPact();
     }
 
     @Test
-    @PactTestFor(pactMethod = "submitCaseWorkerDetails")
-    public void submitForCaseWorker() throws Exception {
+    @PactTestFor(pactMethod = "submitEventForCitizen")
+    public void verifySubmitEventForCitizen() throws Exception {
 
-        caseDataContent = getCaseDataContentWithPath(createEventId, VALID_PAYLOAD_PATH);
+        caseDataContent = PactDslFixtureHelper.getCaseDataContent(HWF_APPLICATION_ACCEPTED);
 
-        CaseDetails caseDetailsReponse = coreCaseDataApi.submitForCaseworker(SOME_AUTHORIZATION_TOKEN,
+        CaseDetails caseDetails = coreCaseDataApi.submitEventForCitizen(SOME_AUTHORIZATION_TOKEN,
             SOME_SERVICE_AUTHORIZATION_TOKEN, USER_ID, jurisdictionId,
-            caseType, true, caseDataContent);
+            caseType,CASE_ID.toString(),true,caseDataContent);
 
-        assertNotNull(caseDetailsReponse);
-        assertNotNull(caseDetailsReponse.getCaseTypeId());
-        assertEquals(caseDetailsReponse.getJurisdiction(), "DIVORCE");
-
-        assertCaseDetails(caseDetailsReponse);
-
+        assertThat(caseDetails.getId(), is(CASE_ID));
+        assertThat(caseDetails.getJurisdiction(), is("DIVORCE"));
+        assertCaseDetails(caseDetails);
     }
 
     @Override
     protected Map<String, Object> getCaseDataContentAsMap(CaseDataContent caseDataContent) throws JSONException {
         Map<String, Object> caseDataContentMap = super.getCaseDataContentAsMap(caseDataContent);
-        caseDataContentMap.put(EVENT_ID, createEventId);
+        caseDataContentMap.put(EVENT_ID, caseDataContent.getEvent().getId());
         return caseDataContentMap;
     }
