+++ src/contractTest/java/uk/gov/hmcts/reform/divorce/casemaintenanceservice/client/ccd/SubmitForCaseWorkerConsumerTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.divorce.casemaintenanceservice.client;
+package uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.ccd;
 
 import au.com.dius.pact.consumer.dsl.PactDslWithProvider;
 import au.com.dius.pact.consumer.junit5.PactTestFor;
@@ -9,29 +9,24 @@ import org.apache.http.client.fluent.Executor;
 import org.json.JSONException;
 import org.junit.After;
 import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
-import uk.gov.hmcts.reform.ccd.client.CoreCaseDataApi;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDataContent;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.ccd.client.model.Event;
-import uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.DivorceCaseMaintenancePact;
-import uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.PactDslFixtureHelper;
+import uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.CcdConsumerTestBase;
 
 import java.util.Map;
 
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
 import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.AssertionHelper.assertCaseDetails;
 import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.ObjectMapperTestUtil.convertObjectToJsonString;
 import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.PactDslBuilderForCaseDetailsList.buildCaseDetailsDsl;
-import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.PactDslFixtureHelper.getCaseDataContent;
 import static uk.gov.hmcts.reform.divorce.casemaintenanceservice.client.util.PactDslFixtureHelper.getCaseDataContentWithPath;
 
-
-public class DivorceCaseMaintenanceSubmitForCitizen extends DivorceCaseMaintenancePact {
+public class SubmitForCaseWorkerConsumerTest extends CcdConsumerTestBase {
 
     private CaseDataContent caseDataContent;
 
@@ -39,6 +34,7 @@ public class DivorceCaseMaintenanceSubmitForCitizen extends DivorceCaseMaintenan
 
     @BeforeAll
     public void setUp() throws Exception {
+
         Thread.sleep(2000);
         caseDataContent = CaseDataContent.builder()
             .eventToken("someEventToken")
@@ -53,17 +49,17 @@ public class DivorceCaseMaintenanceSubmitForCitizen extends DivorceCaseMaintenan
     }
 
     @After
-    public void teardown() {
+    void teardown() {
         Executor.closeIdleConnections();
     }
 
     @Pact(provider = "ccdDataStoreAPI_Cases", consumer = "divorce_caseMaintenanceService")
-    RequestResponsePact submitForCitizen(PactDslWithProvider builder) throws Exception  {
+    RequestResponsePact submitCaseWorkerDetails(PactDslWithProvider builder) throws Exception {
         // @formatter:off
         return builder
-            .given("A Submit for a Citizen is requested", getCaseDataContentAsMap(caseDataContent))
-            .uponReceiving("A Submit For a Citizen")
-            .path("/citizens/"
+            .given("A Submit for a Caseworker is requested" ,  getCaseDataContentAsMap(caseDataContent))
+            .uponReceiving("A Submit For a Caseworker")
+            .path("/caseworkers/"
                 + USER_ID
                 + "/jurisdictions/"
                 + jurisdictionId
@@ -72,10 +68,9 @@ public class DivorceCaseMaintenanceSubmitForCitizen extends DivorceCaseMaintenan
                 + "/cases")
             .query("ignore-warning=true")
             .method("POST")
-            .headers(HttpHeaders.AUTHORIZATION, SOME_AUTHORIZATION_TOKEN, SERVICE_AUTHORIZATION,
-                SOME_SERVICE_AUTHORIZATION_TOKEN)
-            .matchHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
+            .headers(HttpHeaders.AUTHORIZATION, SOME_AUTHORIZATION_TOKEN, SERVICE_AUTHORIZATION, SOME_SERVICE_AUTHORIZATION_TOKEN)
             .body(convertObjectToJsonString(getCaseDataContentWithPath(createEventId, VALID_PAYLOAD_PATH)))
+            .matchHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
             .willRespondWith()
             .status(HttpStatus.SC_CREATED)
             .body(buildCaseDetailsDsl(CASE_ID))
@@ -84,15 +79,21 @@ public class DivorceCaseMaintenanceSubmitForCitizen extends DivorceCaseMaintenan
     }
 
     @Test
-    @PactTestFor(pactMethod = "submitForCitizen")
-    public void verifySubmitEventForCitizen() throws Exception {
+    @PactTestFor(pactMethod = "submitCaseWorkerDetails")
+    public void submitForCaseWorker() throws Exception {
+
         caseDataContent = getCaseDataContentWithPath(createEventId, VALID_PAYLOAD_PATH);
 
-        CaseDetails caseDetailsReponse = coreCaseDataApi.submitForCitizen(SOME_AUTHORIZATION_TOKEN,
+        CaseDetails caseDetailsReponse = coreCaseDataApi.submitForCaseworker(SOME_AUTHORIZATION_TOKEN,
             SOME_SERVICE_AUTHORIZATION_TOKEN, USER_ID, jurisdictionId,
-            caseType,true,caseDataContent);
+            caseType, true, caseDataContent);
+
+        assertNotNull(caseDetailsReponse);
+        assertNotNull(caseDetailsReponse.getCaseTypeId());
+        assertEquals(caseDetailsReponse.getJurisdiction(), "DIVORCE");
 
         assertCaseDetails(caseDetailsReponse);
+
     }
 
     @Override
